Restructured the /move endpoint
