+++ player/router.js
@@ -20,25 +20,24 @@ function factory(stream) {
     }
   });
 
-  router.post("/player", async (req, res, next) => {
+  // router.post("/player", async (req, res, next) => {
+  //   try {
+  //     const { userId } = toData(req.body.jwt)
+  //     const player = await Player.create({ userId, gameId: req.body.gameId });
+  //     const updatedGame = await Game.findByPk(req.body.gameId, { include: [{ model: User, attributes: ['id', 'name'] }] })
 
-    try {
-      const { userId } = toData(req.body.jwt)
-      const player = await Player.create({ userId, gameId: req.body.gameId });  //userId, gameId:req.body.gameId}
-      const updatedGame = await Game.findByPk(req.body.gameId, { include: [{ model: User, attributes: ['id', 'name'] }] })
-
-      const action = {
-        type: "UPDATE_GAME",
-        payload: updatedGame
-      }
-      const string = JSON.stringify(action)
-      stream.send(string)
-      res.send(player);
-    } catch (error) {
-      next(error);
-    }
-   
-  })
+  //     const action = {
+  //       type: "UPDATE_GAME",
+  //       payload: updatedGame
+  //     }
+  //     const string = JSON.stringify(action)
+  //     stream.send(string)
+  //     res.send(player);
+  //   } catch (error) {
+  //     next(error);
+  //   }
+  // }
+  // )
 
   return router
 }
